import accelerator data manage

+++ internal/types/types.go
@@ -42,10 +42,10 @@ type MeasureIndex struct {
 }
 
 type MeasureAccelerometer struct {
-	Id   string    `json:"id,optional"`
-	Data []float32 `json:"data"`
-	Time []int64   `json:"time"` //time tick
-	List []string  `json:"list"`
+	Id   string   `json:"id,optional"`
+	Data []int64  `json:"data"`
+	Time []int64  `json:"time"` //time tick
+	List []string `json:"list"`
 }
 
 type GoalItem struct {
